+++ src/main/java/com/qa/calculator/Subtract.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Subtract implements Editable {
 	
-	public int editNumbers(int num1, int num2) {
+	public double editNumbers(int num1, int num2) {
 		return num1 - num2;
 	}
 }
